@@ -27,8 +27,8 @@
     "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/common": "^11.2.0",
-    "@angular/core": "^11.2.0",
-    "@angular/elements": "^11.2.0"
+    "@angular/common": "~12.2.10",
+    "@angular/core": "~12.2.10",
+    "@angular/elements": "~12.2.10"
   }
 }
